feat: コードを共通化しました。
