+++ app/src/main/java/jp/panta/misskeyandroidclient/ui/users/viewmodel/FollowFollowerViewModel.kt
@@ -7,17 +7,15 @@ import androidx.lifecycle.viewModelScope
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
-import jp.panta.misskeyandroidclient.util.eventbus.EventBus
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import net.pantasystem.milktea.model.user.FollowFollowerPagingStore
 import net.pantasystem.milktea.model.user.RequestType
-import net.pantasystem.milktea.model.user.User
 
 class FollowFollowerViewModel @AssistedInject constructor(
     followFollowerPagingStoreFactory: FollowFollowerPagingStore.Factory,
     @Assisted val type: RequestType
-) : ViewModel(), ShowUserDetails {
+) : ViewModel() {
 
     companion object
 
@@ -32,7 +30,6 @@ class FollowFollowerViewModel @AssistedInject constructor(
     val users = followFollowerPagingStore.users.asLiveData()
     val state = followFollowerPagingStore.state
 
-    val showUserEventBus = EventBus<User.Id>()
 
 
     fun loadInit() = viewModelScope.launch(Dispatchers.IO) {
@@ -45,11 +42,6 @@ class FollowFollowerViewModel @AssistedInject constructor(
         followFollowerPagingStore.loadPrevious()
     }
 
-
-
-    override fun show(userId: User.Id?) {
-        showUserEventBus.event = userId
-    }
 
 
 }
